project/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.5.0
pypdf>=5.0.0
numpy>=2.0.0
mediapipe>=0.10.0
opencv-python>=4.10.0
pyautogui>=0.9.54
pytest>=8.0.0
httpx>=0.27.0
python-multipart>=0.0.9
boto3>=1.34.0
langchain-aws>=0.1.0
mangum>=0.17.0
watchtower>=3.3.0
